move bottlerocket common types to own crate
Move the simple types use by both testsys and the bottlerocket agents
to a common crate that both can use.
