+++ Dockerfile
@@ -32,8 +32,8 @@ ENV PKG_CONFIG_ALLOW_CROSS=1
 ENV CARGO_HOME=/src/.cargo
 ENV OPENSSL_STATIC=true
 
-# Build bottlerocket-agents
-WORKDIR /src/bottlerocket/agents
+# Build bottlerocket agents
+WORKDIR /src/bottlerocket/agents/
 RUN cp -p /src/LICENSE-APACHE /src/LICENSE-MIT /usr/share/licenses/testsys && \
     /usr/libexec/tools/bottlerocket-license-scan \
       --clarify /src/clarify.toml \
