+++ bottlerocket/types/src/lib.rs
@@ -1,0 +1,5 @@
+/*
+This library is for core types that may be used by testsys, agents, and any other Bottlerocket-
+specific crates.
+ */
+pub mod agent_config;
