+++ testsys/src/run_vmware.rs
@@ -1,9 +1,8 @@
 use crate::error::{self, Result};
-use bottlerocket_agents::sonobuoy::Mode;
-use bottlerocket_agents::wireguard::WIREGUARD_SECRET_NAME;
-use bottlerocket_agents::{
-    MigrationConfig, SonobuoyConfig, TufRepoConfig, VSphereClusterInfo, VSphereVmConfig,
-    AWS_CREDENTIALS_SECRET_NAME, VSPHERE_CREDENTIALS_SECRET_NAME,
+use bottlerocket_types::agent_config::{
+    MigrationConfig, SonobuoyConfig, SonobuoyMode, TufRepoConfig, VSphereClusterInfo,
+    VSphereVmConfig, AWS_CREDENTIALS_SECRET_NAME, VSPHERE_CREDENTIALS_SECRET_NAME,
+    WIREGUARD_SECRET_NAME,
 };
 use kube::ResourceExt;
 use kube::{api::ObjectMeta, Client};
@@ -48,7 +47,7 @@ pub(crate) struct RunVmware {
     /// `certified-conformance`, `quick`. Although the Sonobuoy binary defaults to
     /// `non-disruptive-conformance`, we default to `quick` to make a quick test the most ergonomic.
     #[structopt(long, default_value = "quick")]
-    sonobuoy_mode: Mode,
+    sonobuoy_mode: SonobuoyMode,
 
     /// The kubernetes conformance image used for the sonobuoy test.
     #[structopt(long)]
